Add Browse Notes Fragment to display list of notes.
This change adds a BrowseNotesFragment class and a layout, Fragment_browse_notes.xml, to support
the display of a list of notes that the user has created using the app.

Using a LinearLayoutManager, the RecyclerView, displays a list of the notes, by using a CardView. To support the use of CardView's to display notes, a project dependency is added to  "androidx.cardview:cardview:1.0.0" in the build.grade file for the project.